refactor(sigstore): move HandlerOption to handler_options.go, drop httpClient from constructor
- Move HandlerOption type and With* functions to handler_options.go for
  readability (addressing review request)
- Drop httpClient parameter from NewCosignBinary — constructor always
  defaults; callers override via WithHTTPClient after construction
- Add comment on cosignDownloadURL explaining parametric goos/goarch for
  testability

On-behalf-of: Gerald Morrison (SAP) <gerald.morrison@sap.com>
